docs: move URL for "Cilium-Nightly-Tests Job", fix links for CI doc
Big source files for documentation occasionally get split out into
several pieces. At one occasion, the address for hypertext links on
"Cilium-Nightly-Tests Job" was not moved to the same file as the
reference itself, thus breaking the links. Sphinx apparently complained
and the references were later turned into simple preformatted chunks.

Move the URL to the relevant file (it is unused in its current file
anyway) and restore the links.

Also update the name and target for this link, and the target for the
Flannel job.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>